requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
playwright==1.40.0
jinja2==3.1.2
pydantic==2.4.2
langchain==0.0.350
openai==1.3.7
tiktoken==0.5.2
tenacity==8.2.3 